try moving construct libs to submodules
