+++ src/cdk8s/cdk8s-construct.ts
@@ -1,4 +1,4 @@
-import { ConstructLibrary, ConstructLibraryOptions } from './construct-lib';
+import { ConstructLibrary, ConstructLibraryOptions } from '../constructs/construct-lib';
 
 export interface ConstructLibraryCdk8sOptions extends ConstructLibraryOptions {
   /**
