+++ src/constructs/index.ts
@@ -1,0 +1,2 @@
+export * from './construct-lib';
+export * from './jsii-project';
